do not remove compiled template on bundle_rsync

+++ ext_capistrano/lib/ext_capistrano/capistrano/helpers/file_helper.rb
@@ -30,7 +30,7 @@ module ExtCapistrano
       compile_erb(source)
       upload_file(server, source, destination)
       # TODO remove parent directory if it's empty
-      FileUtils.rm_f source
+      FileUtils.rm_f source unless ENV['BUNDLE_RSYNC']
     end
 
     def compile_erb(source)
